# src/docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Opero API",
    "description": "Auto-generierte OpenAPI-Doku (swagger-autogen)",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/api/products/": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Alle Produkte abrufen",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "enum": {
                  "type": "array",
                  "example": [
                    "food",
                    "goods"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "format": {
                  "type": "string",
                  "example": "uuid"
                }
              }
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Suche in name/sku"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                },
                "minimum": {
                  "type": "number",
                  "example": 1
                },
                "maximum": {
                  "type": "number",
                  "example": 200
                }
              }
            },
            "default": 50
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                },
                "minimum": {
                  "type": "number",
                  "example": 0
                }
              }
            },
            "default": 0
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Neues Produkt anlegen",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "kind"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Pizza Margherita"
                  },
                  "kind": {
                    "type": "string",
                    "enum": [
                      "food",
                      "goods"
                    ],
                    "example": "food"
                  },
                  "price": {
                    "type": "number",
                    "minimum": 0,
                    "example": 12.5
                  },
                  "vat_rate": {
                    "type": "number",
                    "minimum": 0,
                    "example": 2.6
                  },
                  "sku": {
                    "type": "string",
                    "example": "PIZ-MARG-001"
                  },
                  "category_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "metadata": {
                    "type": "object",
                    "example": {
                      "plu": 123,
                      "tags": [
                        "vegi"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{id}": {
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Produkt ändern",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "format": {
                  "type": "string",
                  "example": "uuid"
                }
              }
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number",
                    "minimum": 0
                  },
                  "vat_rate": {
                    "type": "number",
                    "minimum": 0
                  },
                  "kind": {
                    "type": "string",
                    "enum": [
                      "food",
                      "goods"
                    ]
                  },
                  "sku": {
                    "type": "string"
                  },
                  "category_id": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "metadata": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Produkt löschen",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}